OAK-3289 : store longs not integers as RDBMK otherwise auto-returns longs anyway
git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/trunk@1697677 13f79535-47bb-0310-9956-ffa450edef68

+++ oak-core/src/test/java/org/apache/jackrabbit/oak/plugins/document/ClusterViewDocumentTest.java
@@ -144,7 +144,7 @@ public class ClusterViewDocumentTest {
         assertTrue(id != null && id.length() > 0);
         String createdAt = doc.getCreatedAt();
         assertTrue(createdAt != null && createdAt.length() > 0);
-        int createdBy = doc.getCreatedBy();
+        long createdBy = doc.getCreatedBy();
         assertEquals(localClusterId, createdBy);
         assertEquals(1, doc.getViewSeqNum());
         assertEquals(1, doc.getActiveIds().size());
